robust uturn (#1526)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1526

As described in [a Stan forum](https://discourse.mc-stan.org/t/nuts-misses-u-turns-runs-in-circles-until-max-treedepth/9727), the NUTS stopping condition doesn't work properly in some instances.

This diff adds additional u-turn checks between subtrees to prevent this behavior. It adds a u-turn check between the leftmost node of the right tree and the left tree, as well as the rightmost node of the left tree and the right tree.

Reviewed By: mootaz77

Differential Revision: D32954189

fbshipit-source-id: 9f0f41257e232e3eadf330691f0e903213bc77ba