Clarify the logic of prune tree transition ratio
Summary: In the BART algorithm, it is required to calculate the prune_log_transition_ratio in the _prune_log_transition_ratio. This requires the calculation of number of growable leafs in the proposed ("pruned") tree. This diff clarifies the logic for the calculation of number of growable leafs in this tree as a function of the new and old nodes.

Differential Revision: D37808283

fbshipit-source-id: 5118e0b40567a827fd049ee814b217df0c76c2bd