Reword docstring so num_observation refers to length of dataset instead of num_samples
Summary: BART is a tree based Bayesian regression model which fits a covariate matrix, X to estimate the response surface (E[y(0), y(1) | X]). In this diff the docstring was edited for clarity. Previously, the docstrings referred to the length of the dataset (len(X)) as num_samples. However, it was realized that this is confusing with the number of samples along the MCMC dimension. Therefore, we now refer to len(X) as num_observations to avoid this confusion.

Differential Revision: D37820250

fbshipit-source-id: 7002083f9451f787413f0df2d8a158a4dc8a6b0c