Implement support for log prob operator in compiler
Summary:
We need the ability to compile models containing log_prop operators on stochastic distributions for one of Todd's models. We previously added support to BMG for this; we now have support in the compiler.

This is the first time we've supported an operator other than sampling on a stochastic distribution, so it was more than the usual amount of work to add a new operator. Not much more, fortunately.

Reviewed By: gafter

Differential Revision: D37807230

fbshipit-source-id: efd77b3bef79fbfd9a05e7764a6c067d711c483e