Clarify the logic of prune tree transition ratio (#1543)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1543

In the BART algorithm, it is required to calculate the prune_log_transition_ratio in the _prune_log_transition_ratio. This requires the calculation of number of growable leafs in the proposed ("pruned") tree. This diff clarifies the logic for the calculation of number of growable leafs in this tree as a function of the new and old nodes. The previous formulation contained hardcoded value which was opaque and problematic for future maintenance.

Reviewed By: feynmanliang

Differential Revision: D37808283

fbshipit-source-id: f38bee2890b04e70a1e00a8d83bd7df2564ad526