Reword docstring so num_observation refers to length of dataset instead of num_samples (#1544)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1544

BART is a tree based Bayesian regression model which fits a covariate matrix, X to estimate the response surface (E[y(0), y(1) | X]).
In this diff: the docstring was edited for clarity. Previously, the docstrings referred to the length of the dataset (len(X)) as num_samples. However, it was realized that this is confusing with the number of samples along the MCMC dimension. Therefore, we now refer to len(X) as num_observations to avoid this confusion.

Reviewed By: feynmanliang

Differential Revision: D37820250

fbshipit-source-id: aeef9e49db340d7f417f0a1eafad794ffbbd140b