Implement predict_with_intervals (#1536)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1536

Background:
We are building Bayesian Additive Regression Trees (BART) as an experimental causal inference model in Bean
Machine.

In this diff:
We implement a method predict_with_intervals() in the BART API which return not just a prediction but also an interval for each prediction at a desired coverage.

Differential Revision: https://www.internalfb.com/diff/D37729486?entry_point=27

fbshipit-source-id: 003ae31896bd2d1998edb4e61765eb61623fb6ac