Fix predict always uses training data
Summary:
Background:
We are building Bayesian Additive Regression Trees (BART) as an experimental causal inference model in Bean
Machine.

In this diff:
This diff fixes a bug which causes the BART model to always predict on training data. This bug originates from an older iteration where the _predict_step() method was only used during training.
Unittest has been added and the integration test has been modified to detect this.

Differential Revision: D37837114

fbshipit-source-id: 395a7e60696887617bd6bac0a530fd2277bdc9c7