Implement predict_with_intervals (#1536)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1536

Background:
We are building Bayesian Additive Regression Trees (BART) as an experimental causal inference model in Bean
Machine.

In this diff:
We implement a method predict_with_intervals() in the BART API which return not just a prediction but also an interval for each prediction at a desired coverage.

Differential Revision: D37729486

fbshipit-source-id: af4da3bf493f22d9e74feb9222447da51b2867c8