Include a setter for Node children (#1559)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1559

Background

BART and XBART are tree based prediction algorithm which construct ensemble of tree structures to make a prediction. These tree structures are composed of nodes.

In this diff:
We are adding a setter to ```BaseNode```. This is useful for pointwise modifications to ancestors during mutation (for task T124884399) and for typecasting children to ```LeafNode``` during xBART grow-from-root recursion.

Reviewed By: swkarlekar

Differential Revision: D38079458

fbshipit-source-id: 6ecc92ea18e4724a748eb910807c15c449f38524