Add cholesky_ex to list of supported functions in compiler (#1570)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1570

Xitong's latest version of the GEP model uses a different version of the Cholesky operator that was not yet recognized by the compiler. The `cholesky_ex` pytorch API returns a tuple consisting of the result and a tensor containing the indices of the inputs which were invalid inputs.

We have no easy way to replicate this behavior in BMG, so instead we'll have the compiler emulate the pytorch behavior, but assume that there are no errors. It returns a tuple of the graph node and a zero tensor.

Reviewed By: gafter

Differential Revision: D38017818

fbshipit-source-id: e7a010d1e9f493ddb61c4a6c162841b81507117e