Support numpy arrays as constant operands to stochastic operators
Summary:
The compiler crashed when given a numpy array as a "constant" operand to a stochastic operator; we now detect this situation and automatically convert the value to a tensor before building a graph node.

While I was at it, I also fixed a couple of long-standing related problems:

* The error message when using an unsupported operand type in a stochastic operator was terrible; I've fixed that so it is somewhat more clear, though it still could be better. (We could for instance say where the error occurred in the original source code.)

* When a constant is tensor-valued, we now make a copy of the tensor in the accumulated graph. We hope that models will not mutate tensors in-place, but by making a copy we at least protect ourselves against this scenario.

I have not yet considered the scenario where a numpy array is used as an argument to a random variable; that is likely broken in a similar way. I'll follow up on that in a later diff.

Reviewed By: gafter

Differential Revision: D38089671

fbshipit-source-id: b0f2a95173d65228cdd1071477a3354ec995b306