Add num_null_cuts as a hyperparameter for XBART (#1566)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1566

XBART uses a hyper parameter called ```num_null_cuts``` to weight the probability of sampling a "no-split" option (thereby controlling the depth on the trees.

In this diff:
We just introduce the variable in the ```XBART``` API. This variable is not used but we will pass it to the backfitting algorithm when the XBART API calls it.

Differential Revision: https://www.internalfb.com/diff/D38092829?entry_point=27

fbshipit-source-id: bde43f2f5d6c6b6b41cb27643fd3b85cb75b5eab