Make XBART API use grow-from-root (#1574)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1574

XBART is a faster variant of BART which grows a new tree recursively at each MCMC step instead of modifying an existing tree. This is known as the grow from root algorithm. It consists of
Selecting a subset of variables to split on (as opposed to BART which considers all the variables as split candidates).
Sampling a cutpoint (which can also be a stop-split sample) by sampling a variable from the subset and a unique value of that variable as a decision threshhold.
Either stopping the recursion or "growing the node" based on the sampling step in 2.
If stopped, the leaf node mean value is sampled otherwise the children of the node are grown.
The ordering of the covariates along each dimensions (i.e. each variable) are maintained using an index matrix.
In this diff:
We make the XBART API use the grow-from-root algorithm as its MCMC step.
The MCMC step consists of:
1. Growing a new tree
2. Sampling the estimate of noise standard deviation
3. Sampling variable importance weights based on the split counts returned by the grow-from-root step.

Reviewed By: feynmanliang

Differential Revision: D38170413

fbshipit-source-id: 83f7958c2be491835639d5e97552fb767944c358