Add num_null_cuts as a hyperparameter for XBART (#1566)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1566

XBART uses a hyper parameter called ```num_null_cuts``` to weight the probability of sampling a "no-split" option (thereby controlling the depth on the trees.

In this diff:
We just introduce the variable in the ```XBART``` API. This variable is not used but we will pass it to the backfitting algorithm when the XBART API calls it.

Reviewed By: feynmanliang, horizon-blue

Differential Revision: D38092829

fbshipit-source-id: f066f69034cfc4f3b14de83c6ce7c5c66e7da986