Create a test case showing the LROM model does not compile yet
Summary:
This test case is based on Todd's Linear Regression Outliers Marginalized Model notebook. This model does not yet compile because we do not yet support compiling the Unit distribution; this test also shows that we generate a bad crash error message when that happens rather than a clear message.

In an upcoming diff I'll start by fixing the error message.

Differential Revision: D38218995

fbshipit-source-id: ba9369e7c19039e916b2945b7845685ed96ad26f