Create a test case showing the LROM model does not compile yet (#1578)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1578

This test case is based on Todd's Linear Regression Outliers Marginalized Model notebook. This model does not yet compile because we do not yet support compiling the Unit distribution; this test also shows that we generate a bad crash error message when that happens rather than a clear message.

In an upcoming diff I'll start by fixing the error message.

Reviewed By: ToddSmall

Differential Revision: D38218995

fbshipit-source-id: 8d7cf91b6e42f5011e5ad77fe69352b53b4ced31