Fix error message when an unsupported non-torch distribution is used
Summary:
All functions -- including type constructors -- in the torch modules are treated as "special" by the compiler. For special functions we have a "can have stochastic arguments" allow list; if your function call has stochastic arguments and is a special function, but not on the allowed list, then you get an error message saying that the function is not supported by BMG.

The idea here is that user-supplied functions we will attempt to rewrite, and models which have calls to third party libraries we will just pass the graph node and hope for the best, but torch functions that we do not support we *strongly believe* will not work, and so we should stop compilation and let the user know.

The case I missed here was: constructors for distributions that *inherit* from torch.Distribution but are not *declared* in torch.distributions.  Such as our Unit distribution.

We now treat any constructor of a distribution as a special function and give an appropriate error message if an attempt is made to call an unsupported one with stochastic arguments.

As noted in the code, we do not yet treat instance methods of such distributions as special. We should. That will be in a later diff.

Reviewed By: ToddSmall

Differential Revision: D38224341

fbshipit-source-id: fdf55982200f26d18c8a534ed27909248973b9f0