Add test cases showing that vectorized LROM model fails to compile (#1585)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1585

A "Bernoulli trick" version of the LROM model should compile successfully, but it does not because the devectorizer does not know how to devectorize a log_prob operator.

Unfortunately, this illustrates a flaw in the design of the devectorizer: when I implemented it, I assumed that the ONLY operator that would ever have a vectorized distribution as its operand was "sample".  "log_prob" has the novel property that it might have a vectorized distribution as its left operand AND a vectorized value as its right operand.

It is therefore not simply a matter of adding the log_prob operator to the list of devectorizable operators. We're going to have to come up with a more general algorithm that can handle operators whose operands are distributions.

I've added two test cases. One shows a simplified "Bernoulli trick" version of Todd's LROM model, and how it fails today. The other is an even more simplified version which just has the offending log prob operator.

Reviewed By: horizon-blue

Differential Revision: D38229153

fbshipit-source-id: fcf927b17ff8e27c95850744ef4a15a804683b62