Updated BMG Statistics Module
Summary:
This diffs includes the codes for the statistics gathering and generation module for BMG. It contains:
Extensions to graph.h to include an object called Statistics as part of the graph and all the necessary declarations for the Statistics object. The object implements the statistics gathering module in the file  graph_stats.cpp which contains the object implementation of the statistics gathering module.
The test file for the modules, in graph_stat_test.cpp.
The statistics module is documented in https://docs.google.com/document/d/1xwO4CqsqS1ekhFqvUJ-mmJR0hct7H_J7jv4W9rffha8/edit?usp=sharing

Differential Revision: D38283045

fbshipit-source-id: fa5f7549f2d6ab3d4bd35504db0335bc450f717a