Fix prior weighting error in XBART paper
Summary:
Background:
We are developing Accelerated Bayesian Additive Regression Trees (XBART) whic uses a grow-from-root tree sampler to from an ensemble of trees which predict the response surface.

In this diff:
In the XBART paper (https://proceedings.mlr.press/v89/he19a.html), the prior weighting of the no-split termination condition is incorrect. In addition to the expression for kappa(0) given in section 2.1, there is an additional weight term multiplying it. This is the number of cut-points considered for splitting. Without this weighting term, the prior of no-splitting does not match the prior of a node being terminal in Eq.7 of the original BART paper ( https://projecteuclid.org/journals/annals-of-applied-statistics/volume-4/issue-1/BART-Bayesian-additive-regression-trees/10.1214/09-AOAS285.full). This was fixed in a later publication (see Algorithm 5 Line 8, https://knowledge.uchicago.edu/record/2324?ln=en) and has also been fixed in the author's code base (https://github.com/JingyuHe/XBART/blob/2e16523014b413408a39320f7d342883eb8efeef/src/tree.cpp#L1645) but not in the original publication.

This reweighting of the no-split prior leads to shallower trees and MUCH faster runtimes (~5x faster) with no effect on the performance.

Differential Revision: D38558674

fbshipit-source-id: a4c33375484cb40405050c55220ad2da6fb3f1d2