Transitioning to the New Profiler (Step 1) (#1600)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1600

The new profiler code has been tested but not enabled. In this Diff and the few following ones, the transition to the new code will take place. In this one, we eliminate the file new_profiler.h and rename the test files to simply profiler.h. This was recommended by both Lily Li and Eric Clippert in the original submission of the new profiler code.

Reviewed By: horizon-blue

Differential Revision: D38480357

fbshipit-source-id: 7916f0569bf33e5b1a49fd7b323922e6bd2b98c5