Deprecate VerboseLevel (#1603)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1603

Per our earlier discussion on the [ergonomics of VerboseLevel in Workplace a long while ago](https://fb.workplace.com/groups/beanmachinesupport/posts/1105996589945318/?comment_id=1106012913277019&reply_comment_id=1106071086604535), it's quite cumbersome to having to use a specialized `Enum` object when there's only a binary option. [We've also seen people trying to pass a boolean to the argument but was confused when this didn't work out](https://fburl.com/anp/aphy6cn0). In addition to this, `verbose` is also a bit of misnomer here as all this flag control the display of the progress bar.

In this diff, we deprecate the use of `VerboseLevel` and the `verbose` argument in `infer`. Moving forward, users should use the new boolean argument `progress_bar` to control whether the tqdm progress bar is displayed or not.

Reviewed By: CactusWin, ToddSmall

Differential Revision: D38295920

fbshipit-source-id: 0399cc0ed69fc7ccfd05539afea50da49905711e