Remove NeuTra from experimental
Summary:
It's been 2 years since the NeuTra module was initially implemented but there's no active use case for it. It's time for us to clean it out to reduce our tech debt. Note that we always have the option to retrieve the code from commit history in the future if needed.

deleteinterncode

Differential Revision: D38927993

fbshipit-source-id: 72da09189a780297d4f6a5895bca6f1d4c0d4b27