Remove NeuTra from experimental (#1614)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1614

It's been 2 years since the NeuTra module was initially implemented but there's no active use case for it. It's time for us to clean it out to reduce our tech debt. Note that we always have the option to retrieve the code from commit history in the future if needed.

deleteinterncode

Reviewed By: feynmanliang

Differential Revision: D38927993

fbshipit-source-id: c84a6add1c637f1e46c3d7bb7fbf2aeae9ea5cef