Eliminate dead code during marginalization rewrite
Summary:
After the transformation enabled by the flag `_fix_observe_true = True`, the Beanstalk compiler produces a graph with some dead nodes. Those nodes cause NUTS in BMG to fail.  If we remove that dead code, we we are one step closer to being able to handle those models.

This diff takes the approach of removing only the specific dead code resulting from that particular transformation.  In the long run we should have a more general pass, either in Beanstalk or in BMG, that removes dead code.

Reviewed By: ericlippert

Differential Revision: D38433962

fbshipit-source-id: 485e479c879b2055f82e8764701f135195cdfe30