copy devectorize-transformer
Summary:
devectorizes nodes that are not supported as tensor operations. It makes no promises that the overhead of read/write operations necessary to split and merge tensors does not outweigh the benefit of running the model in the C++ backend because it does not know or attempt to track memory layouts.

Also there is no protection against graph blowup. We need more tensor operations (complement, minus, log, etc)

Reviewed By: ericlippert

Differential Revision: D38907450

fbshipit-source-id: bdfb82d5f4eb7413d4d22473a54d4736d7b4b7e4