Remove the try-except block in evaluating log prob of a variable (#1635)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1635

We've been silently converting `RuntimeError` and `ValueError` to `-inf` when evaluating log prob of random variable, which was initially design to allow open universe model with change in support, but it ended up causing us quite a lot of trouble when we want to debug a model.

Thus, this diff removes the try-except block in the definition of `Variable` to make debugging easier.

It might sound like that we are going to have a regression in terms of models we can support, but the try-except didn't really fix anything either, because a log prob of `-inf` will lead to a rejection in MH step, so even if we aren't throwing the error, the inference will still get stuck.

Reviewed By: CactusWin

Differential Revision: D39115109

fbshipit-source-id: 09fc4653c1a89f4a141935465866de9e320b1179