Warn user against using torch tensors as arguments of random variables (#1639)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1639

Bean Machine uses the hash value of the arguments of random variables to identify them, which means that foo(1) and foo(tensor(1)) are considered two different random variables. Further in PyTorch, tensors are hashed by memory address instead of by value, so we can have hash(tensor(1)) != hash(tensor(1)). Therefore, it’s not recommended to use tensors as indices of random variables.
In this change,
1. In `rv_identifier.py` we identify if tensors are used as arguments to RVs and warn the user against its use.
2. Added a test case to `rv_identifier_test.py` to check if the warning is triggered correctly when the user provides a tensor instead of a primitive argument.

Reviewed By: horizon-blue

Differential Revision: D39169577

fbshipit-source-id: b37aa6eb4b8878305ada02d9fb7d6aebb1c71400