Fixed incorrect typehints for HME infer() method
Summary:
Was running into some typehint errors with the outputs of model.infer() in D38816344 and realized that the typehint for infer() was wrong.

Tuple[pd.DataFrame] implies a tuple length of 1, however the tuple length here is 2.

Changed it to
` -> Tuple[pd.DataFrame, pd.DataFrame]` in the corresponding classes.

Reviewed By: SepehrAkhavan

Differential Revision: D39195904

fbshipit-source-id: ae249331acabf24a2c5c7fe5d0519f1211a04d1c