Refactor numeric classes to have operations defined outside the class (#1643)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1643

Previously, operators such as `operator+` were defined as members of the class on which they operate.  This moves them to the namespace scope, as recommended by many writers about proper C++ usage.  In addition, the order of the parameters of polygamma was changed.  It used to be `x.polygamma(n)`.  Now it is `polygamma(n, x)`.

I'm sorry for giving such a large diff, but it is all systematic refactoring with no change of behavior intended.

Reviewed By: feynmanliang

Differential Revision: D39173709

fbshipit-source-id: 24e720da036e0de2e610e6799a1bbdf7d6175452