Add compiler support for Poisson distribution
Summary:
We have support for Poisson distribution in BMG. A bunch of the steps for this have already been done. In this diff I complete support for the same in the compiler and added tests for the functionality.

Specifically, we check the graphs produced by Beanstalk in the following cases:
1. [correct] RV with poisson distribution with constant rate.
2. [correct] RV with poisson distribution, where the rate has a value sampled from a positive real RV.
3. [correct] RV with poisson distribution, with tensor rate.
4. [error] RV with poisson distribution, where the rate has a value samples from a negative real RV.

Differential Revision: D39286270

fbshipit-source-id: 254049b4e024dffc24a29fc2e481d25948f9f3b0