Throw a meaningful error when user tries to query a discrete random variable with HMC/NUTS (#1651)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1651

HMC algorithms (including its variants NUTS) can only perform inference on continuous latent random variable, and should raise an error if user try to query a discrete latent random variable. Currently, Bean Machine doesn't perform an explicit check for the support of the distribution, and may either throw an obscured error message during part of inference (e.g. LongTensor doesn't support gradient) or silently return an incorrect inference result (e.g. when a model contains a Poisson distribution).

In this diff, I check whether a discrete latent random variable is present in the initialization of `RealSpaceTransform` and raise an error if it is the case. Added a test case in `hmc_utils_test.py` to test this new addition.

Reviewed By: horizon-blue

Differential Revision: D39314722

fbshipit-source-id: ffcdc45928f2c197e99d9c9fb0cbaf022da1851e