Separate computation of deterministic and stochastic ancestors from Graph implementation. (#1664)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1664

Class Graph was keeping track of "deterministic and stochastic ancestors" as fields, complicating the class and especially the `add_node` method. This was unnecessary especially since this data is only currently being used by the experimental (not used?) CAVI inference method.
This functionality is now separate from class Graph and implemented as a regular function in `support.cpp`. This greatly simplifies `Graph::add_node`, which in turn will make it easier to implement a new method `Graph::remove_node`, the original motivation for this diff.

Differential Revision: D39449301

fbshipit-source-id: 112703c21172238aa0d49fa9a8c7fb74655a7842