Fix nightly Hierarchical Modeling tutorial test
Summary: The nightly test for the Hierarchical Modeling tutorial [has been failing because we hard-coded the reshape dimension to `(4, 5000)` in tutorial utils](https://www.internalfb.com/intern/sandcastle/job/13510799651396224/), which does't match the dimension of the samples we get from running smoke test. This can be fixed by updating the utils to determine the shape by checking the `MonteCarloSamples` object.

Differential Revision: D39495458

fbshipit-source-id: 93b9811e6af7f88385cbdeb94c7263f9090089f7