replace forward gradient tests for MATRIX_SCALE (#1668)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1668

The previous tests for MATRIX_SCALE tested only a subset of the forward derivative results.  It therefore didn't catch the fact that the result was not correct before (it has since been fixed).  This diff replaces the old test with a direct test for all of the the gradients of MATRIX_SCALE.

Reviewed By: shannyang

Differential Revision: D39490728

fbshipit-source-id: c276556b464e83b5b416dc26c549efa7dfb1af72