Make RVIdentifier sortable (for now) (#1667)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1667

As reported in [#1665](https://github.com/facebookresearch/beanmachine/issues/1665), when trying to compile with NNC, functorch (v2.1.0) will attempt to sort [the keys of a dictionary before flattening it](https://github.com/pytorch/functorch/blob/v0.2.1/functorch/_src/aot_autograd.py#L42), where it will raise an exception because Bean Machine's RVIdentifier type is not sortable.

This sorting of dictionary keys is removed in the latest functorch, though until then, as a workaround, we can make `RVIdentifier` sortable for now by defining a `__lt__` operator for it.

Reviewed By: feynmanliang

Differential Revision: D39486353

fbshipit-source-id: 26073eafac1398f2957f70d88e82eaa6cd704f11