Fix nightly Hierarchical Modeling tutorial test (#1670)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1670

The nightly test for the Hierarchical Modeling tutorial [has been failing because we hard-coded the reshape dimension to `(4, 5000)` in tutorial utils](https://www.internalfb.com/intern/sandcastle/job/13510799651396224/), which does't match the dimension of the samples we get from running smoke test. This can be fixed by updating the utils to determine the shape by checking the `MonteCarloSamples` object.

Reviewed By: CactusWin

Differential Revision: D39495458

fbshipit-source-id: fe11debb9122c0acb4f4f65aa9d17947e8003b87