Fix data loading in several of the nightly tutorial tests
Summary: A few of our tutorial use the data loading utils, which attempt to create a directory to cache the downloaded data. However, it appears that sandcastle vm does not permit `mkdir`, so they all fails with `OSError: Function not implemented`. Since the caching of the data is completely optional, we can skip this step if anything error occurs, and the rest of tutorial should run just fine.

Differential Revision: D39517456

fbshipit-source-id: 79653cb1687c1a6b029c602f537af05d15ffb2c9