Automate generation of Graph's cached structural information. (#1671)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1671

Graph has a number of auxiliary structures: which nodes are stochastic, which ones affect which others, and so on.
Previously, client code that to explicit ask these to be computed, which is error-prone.
This diff makes this data obtained through getters only, which calculate them as needed if they have not been calculated yet. Client code no longer needs to be concerned about that (see for example changes in `mh.cpp` and `global_state.cpp`).
This change was further prompted by the fact that now one can remove nodes from the graph. This invalidates the auxiliary data structures. Instead of hoping for the user to realize that and request their recalculation, it is better to invalidate the cache and have it automatically computed if need be.
For good measure, `add_node` now likewise invalidates the cache, since it also renders the information obsolete.

Reviewed By: gafter

Differential Revision: D39515797

fbshipit-source-id: 09ae73ea05a07cfc8a9bee94b7e4168eddb87bc1