Separate one huge header minibmg into many more specific headers.
Summary:
Reorganizes the source code of minibmg into a bunch of separate headers and source files.  There are no meaningful changes to the actual lines of code.  They are just moved around.

No functional changes.  We keep `minibmg.h` as a header to pull in the other headers for clients.

Differential Revision: D39588208

fbshipit-source-id: 6ceb403021a52af68c3659189a39e3b02850a37c