Add consistently named methods for adding constants to BMG. (#1677)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1677

For historical reasons BMG has a bunch of different ways to add a constant single value:

    uint add_constant(bool value);
    uint add_constant(double value);
    uint add_constant(natural_t value);
    uint add_constant(NodeValue value);
    uint add_constant_probability(double value);
    uint add_constant_pos_real(double value);
    uint add_constant_neg_real(double value);

Notice that the first four are overloads and the rest describe the semantics of the value added.

This then leads to some unfortunate confusion and ambiguities. Notice how all our test cases which I've updated said stuff like:

    add_constant((graph::natural_t)2)

to ensure that we add a natural rather than a real valued constant. If you're having to always insert casts to make overload resolution clear, that's evidence for the proposition that using overloads is a bad idea.

We will solve this problem in two stages, to prevent breaking changes. First, I'll add three named methods:

    uint add_constant_bool(bool value);
    uint add_constant_real(double value);
    uint add_constant_natural(natural_t value);

that are unambiguous. In a later diff I will remove the now-deprecated add_constant overrides.

Reviewed By: gafter

Differential Revision: D39585600

fbshipit-source-id: 6b2dbc67b94c577de053c0d49649347976c26676