Fix compiler's C++ codegen, part one
Summary:
I never actually attempted to compile any of the code fragments that the BeanMachine-to-C++ compiler outputs and so naturally a few small syntax errors crept in. Thanks to Neal for listing them for me. This diff fixes the ones Neal found; there are a couple more that will come in a later diff. The fixes are:

* declaration of matrix variables was missing a semi
* we now use the unambiguous add_constant_blah methods created by the previous diff
* fixed the cut-n-paste error that led to brackets in observation argument lists

Reviewed By: gafter

Differential Revision: D39526801

fbshipit-source-id: 41d81a7b77a004d8e25fcba3b4fe69dc2f06385f