Fix data loading in several of the nightly tutorial tests (#1672)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1672

A few of our tutorial use the data loading utils, which attempt to create a directory to cache the downloaded data. However, it appears that sandcastle vm does not permit `mkdir`, so they all fails with `OSError: Function not implemented`. Since the caching of the data is completely optional, we can skip this step if anything error occurs, and the rest of tutorial should run just fine.

Reviewed By: gafter

Differential Revision: D39517456

fbshipit-source-id: ac3f52eb2ba20bd077cef6b78cb7957041a4ddbf