matrix_complement in beanstalk (#1682)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1682

We have recently added `MATRIX_COMPLEMENT`to BMG; in this diff I add support in the compiler for the same.
Additional changes apart from the usual changes to add a node include,
 1. changes to `_type_meets_requirement` in `fix_requirements.py` to handle, boolean, probability and simplex matrix.
 2. changes to `_constant_factories` in `copy_and_replace.py` to allow for constantprobabilitymatrix.
 3. changes to `_tensor_constants` in `devectorizer_transformer.py` to support constant matrix operators as inputs to matrix_complement.

This diff works for the default case when a matrix is passed to a matrix_compelement node or when untensorized complements are part of the original graph . However, when complements are produced as a result of say `arithmetic_graph_fixer`, it does not tensorize a set of complements into a tensor. This is due to the order of execution of different `node_fixers`.

For example, when the original graph contains `log1p(1-beta())`, in this case the arithmetic_fixer converts it into `log(complement(beta())`. But this happens after `vectorized_graph_fixer`, therefore we do not tensorize these nodes.
I will investigate and address this in an upcoming diff.

Reviewed By: ericlippert

Differential Revision: D39652593

fbshipit-source-id: 41427a205d6f61df2923b98cf709afff49cafa5e