Expose flag to discard burn in samples in beanstalk BMGInference (#1675)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1675

In some cases, it is useful to discard the first few samples of inference to get better chain mixing, therefore improved `r_hat` and `ess` metrics.
`MonteCarloSamples` already has logic implemented to discard the specified number of samples using the optional `num_adaptive_samples` parameter.
In this diff, I expose this parameter one level up in `BMGInference.infer()`.
Added a test case to ensure that it works correctly in the default and non-default case.

Reviewed By: ericlippert, horizon-blue

Differential Revision: D39565178

fbshipit-source-id: 33cbecd6dd245c99630a494b53382eaaaa0977d5