Fix the end of adaptation check
Summary:
Just realized that there's a serious typo in the implementation of `Sampler`. As a result, it will never call `proposer.finish_adaptation()` (unless `num_samples=0`, which is rarely the case).

I suspect this is what causes the big variance in run time in some of the HMC/NUTS runs. In particular, without calling `finish_adaptation()`, we will *not* be setting the step size to the dual-averaged value. Instead, we will be using the step size learned from the last [adaptation window](https://mc-stan.org/docs/reference-manual/hmc-algorithm-parameters.html), which is only 50 iterations wide.

Differential Revision: D39707018

fbshipit-source-id: bffc55f43e4ccc6af56cefa67385dc2b5d0ffe52