Fix the end of adaptation check (#1692)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1692

Just realized that there's a serious typo in the implementation of `Sampler`. As a result, it will never call `proposer.finish_adaptation()` (unless `num_samples=0`, which is rarely the case).

I suspect this is what causes the big variance in run time in some of the HMC/NUTS runs. In particular, without calling `finish_adaptation()`, we will *not* be setting the step size to the dual-averaged value. Instead, we will be using the step size learned from the last [adaptation window](https://mc-stan.org/docs/reference-manual/hmc-algorithm-parameters.html#automatic-parameter-tuning), which is only 50 iterations wide.

Reviewed By: CactusWin

Differential Revision: D39707018

fbshipit-source-id: 5f903424a1fdffb64aea2d4a771d586678034c1f