minibmg: unify traced and node.
Summary:
We used to have two different expression DAG impleemntations: the graph nodes, and `Traced`, which was used for tracing and emitting code for automatic derivatives (AD).  There was no reason for them to be separate, other than the fact that the traced version supports operator overloading.  However, separating them required a number of small changes to make them similar enough to merge as a last step.  This is that last step.

This diff replaces the old traced implementation with a trivial wrapper around a node.

Next will be a fluent graph factory that takes advantage of this.

Differential Revision: https://internalfb.com/D39698758

fbshipit-source-id: bf86f5198d5993ad8eff0f9e82e3e2e585bad923