Ensure `graph_test.cpp` test meant to test all types of nodes does indeed do that. (#1696)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1696

In the course of working with `graph_test.cpp` I realized a test requiring a graph of all node types was out-of-date and the graph did not include many types of nodes defined since then.

This diff introduces a new test in the same file checking that this tested graph does contain nodes of all types.

To automate the test and make it robust to future definitions, the test uses an "enum iterable" defined in `util.h` and used for defining such iterables for enums `DistributionType`, `OperatorType` and `FactorType`.

Of course, the diff also needed to include all the definitions of the missing node types in the tested graph.

Reviewed By: gafter

Differential Revision: D39735415

fbshipit-source-id: 86de496a3b299be85e59a456bfc2c81473988c72