Add support for using NUTS inference from BMGInference (#1701)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1701

We want to support NUTS inference via `BMGInference`. In this diff I update `graph.cpp` to expose NUTS api, this is similar to how NMC exposes its API. We make use of `InferConfig` object for both NUTS and NMC to communicate number of warm up samples.

Note, Profiler in `graph.cpp` seems to be specialized to `NMC`, which means the profiling information for NUTS might be incorrect and this may need to be updated. I have created T132701053 to keep track of this.

Reviewed By: gafter

Differential Revision: D39767773

fbshipit-source-id: b614f0801fd637f92977bbe6d3915b19df913628