minibmg: reverse mode automatic differentiation (#1703)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1703

Implement reverse-mode automatic differentiation (AD) for scalars.  Reverse-mode is needed for Variational Inference (VI) and multi-site Markov Chain Monte Carlo (MCMC) methods such as Hamiltonian Monte Carlo (HMC) and the No U-Turn Sampler (NUTS).

Although this is probably a high-performance implementation, the performance of reverse-mode will not likely be on the critical path of inference, as our intent is to use this implementation `Reverse` on top of symbolic nodes `Traced` - i.e. `Reverse<Traced>` - to compute gradients not numerically, but symbolically. That will produce an expression DAG which we can then optimize and generate efficient native code from.  That tightly-coded native code is what would be invoked during inference to compute the gradients (without even needing to traverse the graph).  This is a stepping-stone in that direction.

Reviewed By: shannyang

Differential Revision: D39749431

fbshipit-source-id: 7b13ae924a3bdb4777c1cf92276609980e54acab