Relocate test suite: `src/*/tests/` -> `tests/` (#1674)
Summary:
### Motivation
Currently, the Bean Machine test suite is aligned with and contained inside the library source tree, mostly residing in independent `src/*/tests/` subfolders. To allow for more code sharing and more structural flexibility in the test suite going forward, it is desirable to gather the tests into a dedicated top-level directory. This is in agreement with the default package structure in the official Python documentation [1], and with the Pytest recommendation to maintain tests "outside", rather than "as part of the application code", when a project has "many functional tests" [2].

[1] https://packaging.python.org/en/latest/tutorials/packaging-projects/
[2] https://docs.pytest.org/en/latest/explanation/goodpractices.html#choosing-a-test-layout-import-rules

### Changes proposed
As a first step, this PR moves most tests into the `tests/` directory. Currently excluded from this relocation are:
- `beanmachine/ppl/testlib` (utilities)
- ~~`beanmachine/ppl/experimental/*/tests/*_test.py` (distinct library layer)~~ (note: experimental tests are moved as part of this diff as well)
- `beanmachine/applications/*/tests/*_test.py` (distinct library layer)
- `beanmachine/graph/*/tests/*_test.cpp` (distinct test infrastructure)

The PR is structured into the following commits:
1. relocate Pytest suite
2. relocate GTest suite
3. revert (2) to simplify merging, but keep (2) for reference.

Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1674

Test Plan:
The test coverage and results from `pytest` remain unchanged after the diff.

### Types of changes
- [x] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

### Checklist
- [x] My code follows the code style of this project.
- [x] My change requires a change to the documentation.
- [x] I have updated the documentation accordingly.
- [x] I have read the **[CONTRIBUTING](https://github.com/facebookresearch/beanmachine/blob/main/CONTRIBUTING.md)** document.
- [x] I have added tests to cover my changes.
- [x] All new and existing tests passed.
- [x] The title of my pull request is a short description of the requested changes.

Reviewed By: feynmanliang

Differential Revision: D39562696

Pulled By: horizon-blue

fbshipit-source-id: 3510be6db4a665b1d34185dee637920ea672e3a6