minibmg: add RVID to the sample operation
Summary:
Soon we will be adding a deduplication facility to minibmg.  Unfortunately, distinct samples from the same distribution appear to be duplicate operations, but because samples are nondeterministic they are actually distinct operations.  To make each sample node distinct, we add a random variable identifier.

Most clients of the graph will have no need to pay attention to the identifier.

Differential Revision: D39888681

fbshipit-source-id: 44de5dbec85132c4bec633091ee5981650322a6c