minibmg: add transforms for distributions (#1707)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1707

For inference methods that do not want to be constrained in the range of values proposed for stochastic variables, we provide a set of transformations that map a distribution onto the full set of real numbers.

References:
Change of Variables in statistics:
https://online.stat.psu.edu/stat414/lesson/22/22.2

Stan:
https://mc-stan.org/docs/2_27/reference-manual/change-of-variables-section.html

This diff implements two transformations: log to map from (0..INF) and sigmoid to map from (0..1).  The log transformation is the usual log function (and exp for the inverse).  The sigmoid transformation is logit (and sigmoid for the inverse).

Reviewed By: feynmanliang

Differential Revision: D39789131

fbshipit-source-id: 4186e66b8020b38957233a57c79e4a32bf6d6163