minibmg: add RVID to the sample operation (#1709)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1709

Soon we will be adding a deduplication facility to minibmg.  Unfortunately, distinct samples from the same distribution appear to be duplicate operations, but because samples are nondeterministic they are actually distinct operations.  To make each sample node distinct, we add a random variable identifier.

Most clients of the graph will have no need to pay attention to the identifier.

Reviewed By: ericlippert

Differential Revision: D39888681

fbshipit-source-id: 0e15c0407564f8a49bdc52492bf3335c895c101a