Add BROADCAST and FILL_MATRIX operators to BMG (#1710)
Summary:
Pull Request resolved: https://github.com/facebookresearch/beanmachine/pull/1710

Broadcasting -- taking a tensor of one shape and replicating its contents until it is another, larger shape -- is a common operation in Bean Machine python models. A simple example would be a distribution where the shapes of the inputs are different:

    random_variable def mu(n): [some scalar distribution]
    random_variable def sigma(): [some scalar distribution]
    random_variable def n():
      m = tensor([mu(1), mu(2)]
      s = sigma()
      return Normal(m, s)

Pytorch automatically "broadcasts" the second argument, duplicating the value to pretend that you wrote

      return Normal(m, tensor([s, s]))

We'll need to be able to represent models that use broadcasting semantics in BMG; this is a step in that direction. In future diffs we will add support to the compiler for detecting broadcasting operations and putting them in the graph.

Broadcasting in PyTorch is a fairly complicated operation because tensors are arbitrary n-dimensional rectangles. In BMG we use the Eigen library, so we are restricted to exactly-2-dimensional arrays; this greatly simplifies broadcasting semantics.  The rules are straightforward:

* Operator FILL_MATRIX takes two *constant* naturals giving the size of the resulting matrix, and a non-matrix value which is replicated into the matrix.

* Operator BROADCAST similarly takes the size, and a matrix value which is expanded to the desired size. The input will typically be a single row or single column, which is then replicated to the desired size.

Reviewed By: gafter

Differential Revision: D39489772

fbshipit-source-id: e88f5ba464abf5e342aabd94d8d0c41e8f296406