minibmg: add value-based equality and hash-code support for nodes.
Summary:
In preparation for deduplicating nodes, we add value-based hash code and equality operations that make `std::unordered_set` and `std::unordered_map` consider equivalent node smart pointers to be equivalent keys.  This makes sense from a user's point of view because in minibmg nodes are immutable.

We can still use the ordered set and map types if we want all distinct nodes to be considered distinct.  That is how `topological_sort` works.

Reviewed By: ericlippert

Differential Revision: D39894198

fbshipit-source-id: 71d0f15e5d17a3a20c8733dd0766bb9f67c6ab72